feature: email after checkout
