+++ package.json
@@ -30,7 +30,9 @@
     "express-session": "^1.15.1",
     "history": "^4.6.3",
     "lodash": "^4.17.5",
+    "mailgen": "^2.0.2",
     "morgan": "^1.8.1",
+    "nodemailer": "^4.6.0",
     "passport": "^0.3.2",
     "passport-google-oauth": "^1.0.0",
     "pg": "^6.1.2",
